+++ Classes/Domain/Form/SetupForm.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * @author Robin 'codeFareith' von den Bergen <robinvonberg@gmx.de>
+ * @copyright (c) 2018 by Robin von den Bergen
+ * @license http://opensource.org/licenses/gpl-license.php GNU Public License
+ * @version 1.0.0
+ *
+ * @link https://github.com/codeFareith/cf_google_authenticator
+ * @see https://www.fareith.de
+ * @see https://typo3.org
+ */
+
+namespace CodeFareith\CfGoogleAuthenticator\Domain\Form;
+
+class SetupForm
+{
+    /*─────────────────────────────────────────────────────────────────────────────*\
+            Constants
+    \*─────────────────────────────────────────────────────────────────────────────*/
+    public const FORM_NAME = 'GoogleAuthenticatorSetupForm';
+
+    /*─────────────────────────────────────────────────────────────────────────────*\
+            Properties
+    \*─────────────────────────────────────────────────────────────────────────────*/
+    /** @var string */
+    protected $secret;
+
+    /** @var string */
+    protected $oneTimePassword;
+
+    /*─────────────────────────────────────────────────────────────────────────────*\
+            Methods
+    \*─────────────────────────────────────────────────────────────────────────────*/
+    public function __construct(string $secret, string $oneTimePassword = '')
+    {
+        $this->secret = $secret;
+        $this->oneTimePassword = $oneTimePassword;
+    }
+
+    public function getSecret(): string
+    {
+        return $this->secret;
+    }
+
+    public function setSecret(string $secret): void
+    {
+        $this->secret = $secret;
+    }
+
+    public function getOneTimePassword(): string
+    {
+        return $this->oneTimePassword;
+    }
+
+    public function setOneTimePassword(string $oneTimePassword): void
+    {
+        $this->oneTimePassword = $oneTimePassword;
+    }
+}